% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneNet.R
\name{GeneNet}
\alias{GeneNet}
\title{GeneNet}
\usage{
GeneNet(data, labels, caseName, controlName, meanFilter = 0.2,
  SDFilter = 0.2, edgeCutoff = 0.004, permutationTimes = 10,
  CPUNumbers = 1, outputPrefix = "MetaDCN", pathwayDatabase,
  silent = FALSE)
}
\arguments{
\item{data}{a list of matrice (studies) with row names as genes and column 
names as samples.}

\item{labels}{a list of vectors to specify the case/control labels for 
each subject corresponidng to the columns in each data matrix}

\item{caseName}{a character string denoting case label}

\item{controlName}{a character string denoting control label}

\item{meanFilter}{a number to specify qunatile cutoff for mean}

\item{SDFilter}{a number to specify qunatile cutoff for SD}

\item{edgeCutoff}{a number to specify qunatile cutoff for correlation to 
to define an edge}

\item{permutationTimes}{a number to specify how many permutations to be 
used}

\item{CPUNumbers}{a number to specify how many CPUs are used for parallel,
must be less than permutationTimes}

\item{outputPrefix}{a character string for output file prefix}

\item{pathwayDatabase}{a list with each element as a vector of 
pathway genes}

\item{silent}{TRUE/FALSE to specify if suppress screen output}
}
\value{
a list and several RData files saved in the current working directory
\item{GeneNet }{A list containing important parameters to pass to SearchBM and MetaDCN functions}
\item{AdjacencyMatrice.RData }{A list of adjacency matrice for case and control in each study in the order of case studies and control studies}
\item{CorrelationMatrice.RData }{A list of correlation matrice for case and control each study in the order of case studies and control studies}
\item{CorrelationMatrice.RData }{A list of correlation matrice for case and control each study in the order of case studies and control studies}
\item{AdjacencyMatricePermutationP.RData }{A list of correlation matrice for case and control in each study in the order of case studies and control studies for permutation P}
}
\description{
Generate correction and adjacency matrice for data and permutation
}
\details{
This function will generate correction and adjacency matrix for data and application
}
\examples{
data(pathwayDatabase)
}
\author{
Li Zhu
}

