% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDCN.R
\name{MetaDCN}
\alias{MetaDCN}
\title{MetaDCN}
\usage{
MetaDCN(GeneNetRes, SearchBMRes, FDRCutoff, w1 = NULL, silent = FALSE)
}
\arguments{
\item{GeneNetRes}{A list from GeneNet function}

\item{SearchBMRes}{A list from SearchBM function}

\item{FDRCutoff}{A number to specify FDR cutoff for basic modules}

\item{w1}{A number chosen from (100, 200, ..., 700) to specify the weight1 
used in objective function (optional). If not specified, w1 from SearchBM 
function will be used (recommended).}

\item{silent}{TRUE/FALSE to specify if suppress screen output}
}
\value{
MetaDNC will return a list and files for Cytoscape.

list containing:
\item{w1}{w1 used}  
\item{Supermodule }{Summary of supermodules}

CytoscapeFiles is a folder containing files for Cytoscape.
}
\description{
Assemble basic modules into supermodules
}
\details{
This function will assemble basic modules detected from SearchBM into 
supermodules.
}
\author{
Li Zhu (liz86@pitt.edu)
}

