% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDCN.R
\name{MetaDCN}
\alias{MetaDCN}
\title{MetaDCN}
\usage{
MetaDCN(GeneNetRes, FDRCutoff, w1 = NULL, silent = FALSE)
}
\arguments{
\item{GeneNetRes}{A list from GeneNet function}

\item{FDRCutoff}{A number to specify FDR cutoff for basic modules}

\item{w1}{A number chosen from (100, 200, ..., 700) to specify the weight1 
used in objective function (optional). If not specified, the w1 with the 
most modules will be chosen (recommended).}

\item{silent}{TRUE/FALSE to specify if suppress screen output}
}
\value{
a list and one zip file for Cytoscape
\item{w1}{w1 used}  
\item{ModuleInCase }{Summary of basic modules in case}
\item{ModuleInConrtol }{Summary of basic modules in control}
\item{Supermodule }{Summary of supermodules}
\item{module_assembly_edge_node_list.zip}{Files for Cytoscape}
}
\description{
Assemble basic modules into supermodules
}
\details{
This function will assemble basic modules found from SearchBM into 
supermodules
}
\examples{
data(pathwayDatabase)
}
\author{
Li Zhu
}

