% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDCN_main.R
\name{MetaDCN}
\alias{MetaDCN}
\title{MetaDCN}
\usage{
MetaDCN(data, labels, caseName, controlName, meanFilter = 0.2,
  SDFilter = 0.2, FDRCutoff = 0.3, edgeCutoff = 0.004, MCSteps = 500,
  jaccardCutoff = 0.8, permutationTimes = 10, repeatTimes = 10,
  outputFigure = TRUE, outputPrefix = "MetaDCN", CPUNumbers = 1,
  pathwayDatabase, silent = FALSE)
}
\arguments{
\item{data}{a list of matrix (study) with row names as genes and column 
names as samples.}

\item{labels}{a list of vectors to specify the case/control labels for 
each subject corresponidng to the columns in each data matrix}

\item{caseName}{is the character string denoting case label}

\item{controlName}{is the character string denoting control label}

\item{meanFilter}{a number to specify qunatile cutoff for mean}

\item{SDFilter}{a number to specify qunatile cutoff for SD}

\item{FDRCutoff}{a number to specify FDR cutoff for basic modules}

\item{edgeCutoff}{a number to specify qunatile cutoff for correlation to 
to define an edge}

\item{MCSteps}{a number to specify the maximum number of MC steps for 
simulated annealing}

\item{jaccardCutoff}{a number to specify the maximum jaccard index allowed}

\item{permutationTimes}{a number to specify how many permutations are used}

\item{repeatTimes}{a number to specify how many repeats of different seed 
modules to be used}

\item{outputFigure}{TRUE/FALSE to specify if figures are generated}

\item{outputPrefix}{a character string for output file prefix}

\item{CPUNumbers}{a number to specify how many CPUs are used for parallel,
must be less than permutationTimes}

\item{pathwayDatabase}{a list with each element as a vector of 
pathway genes}

\item{silent}{TRUE/FALSE to specify if suppress screen output}
}
\value{
several csv files and zip files for figures
}
\description{
Detect meta-analysis differential co-expression network (DCN)
}
\details{
This function will generate basic modules and supermodules
}
\examples{
data(pathwayDatabase)
}
\author{
Li Zhu
}

