% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchBM.R
\name{SearchBM}
\alias{SearchBM}
\title{SearchBM}
\usage{
SearchBM(GeneNetRes, MCSteps = 500, jaccardCutoff = 0.8, repeatTimes = 10,
  outputFigure = TRUE, silent = FALSE)
}
\arguments{
\item{GeneNetRes}{a list from GeneNet function}

\item{MCSteps}{a number to specify the maximum number of MC steps for 
simulated annealing}

\item{jaccardCutoff}{a number to specify the maximum jaccard index allowed 
between basic modules}

\item{repeatTimes}{a number to specify how many repeats of different seeds 
to be used}

\item{outputFigure}{TRUE/FALSE to specify if figures are generated}

\item{silent}{TRUE/FALSE to specify if suppress screen output}
}
\value{
Several Rdata, csv and png files saved in the current working directory
\item{outputPrefix_permutation_energy_direction_p.Rdata}{A list of energies of basic modules from permutation p}
\item{outputPrefix_basic_modules_summary_direction_weight_w.csv }{A summary of basic modules found using weight w in forward/backward search}
\item{outputPrefix_threshold_direction.csv }{A table listing number of basic modules found under different FDRs in forward/backward search}
\item{outputPrefix_figure_basic_module_c_repeat_r_direction_weight_w.png }{Plot of basic module from component c from repeat r using weight w in forward/backward search}
}
\description{
Search for differentially co-expressed basic modules
}
\details{
This function will search for basic modules differentially co-expressed 
between case and control
}
\examples{
data(pathwayDatabase)
}
\author{
Li Zhu
}

